part

+++ app/src/main/java/com/example/onlinestore/domain/models/TopProduct.kt
@@ -3,6 +3,6 @@ package com.example.onlinestore.domain.models
 data class TopProduct(
     val id : String,
     val title : String,
-    val prise : Float,
+    val prise : Int,
     val imageUrl :  String,
 )
